gpu offload: address PR #24068 review feedback (round 2)
- inspect.go: Fix txn leak in parseTableTarget — add defer Rollback
  so error paths don't leave transactions open
- manifest_http.go: Distinguish 400 (invalid input) vs 404 (not found)
  status codes in manifest endpoint error responses
- gpu_offload.go: Precompute column types array in buildGPUResultSet
  to avoid per-row gpuTypeToMysql calls

+++ pkg/vm/engine/tae/rpc/manifest_test.go
@@ -77,7 +77,7 @@ func TestGenerateManifest_Basic(t *testing.T) {
 	// No data inserted, so no objects
 	assert.Equal(t, 0, len(m.Objects))
 	assert.NotNil(t, m.Stats)
-	assert.Equal(t, 0, m.Stats.TotalRows)
+	assert.Equal(t, int64(0), m.Stats.TotalRows)
 
 	t.Logf("Manifest JSON:\n%s", string(data))
 }
